Only the render remains in the main
